Renamed csp can driver file.
The csp can driver "can.h" conflicts with the Iris software can driver
"can.h"
Even with the libcsp files exclueded from the softconsole build it still
uses them. Not sure how to have it ignore the file completely.

So the solution is just to rename the csp can driver.